minor refactor: move has_non_science_pixels()
